Example use of facebook's async loading code
Example use of facebook's async loading code instead of Requirements:: adding the script.

This would need further modification to add custom facebook scripts somewhere, as they need to be within the async function call. Alternatively, could couple to jQuery implementations and use $.getScript() with a callback as that's also async.

To be honest, this method should probably be removed completely and handled exclusively in js.